// frontend/src/api/generated/bookings/bookings.zod.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Prototype API
 * OpenAPI spec version: 1
 */
import * as zod from 'zod';




export const createBookingBodyAppointmentTimeRegExp = new RegExp('^([01]\\d|2[0-3]):[0-5]\\d$');
export const createBookingBodyCustomerNameMax = 100;

export const createBookingBodyLicencePlateMax = 15;



export const CreateBookingBody = zod.object({
  "appointmentDate": zod.iso.date(),
  "appointmentTime": zod.string().min(1).regex(createBookingBodyAppointmentTimeRegExp),
  "customerName": zod.string().min(1).max(createBookingBodyCustomerNameMax),
  "licencePlate": zod.string().min(1).max(createBookingBodyLicencePlateMax),
  "washProgramme": zod.enum(['BASIC', 'COMFORT', 'PREMIUM'])
})

export const CreateBookingResponse = zod.object({
  "appointmentDate": zod.iso.date(),
  "appointmentTime": zod.string(),
  "createdAt": zod.iso.datetime({"offset":true}),
  "customerName": zod.string(),
  "id": zod.uuid(),
  "licencePlate": zod.string(),
  "washProgramme": zod.enum(['BASIC', 'COMFORT', 'PREMIUM'])
})

export const GetBookingParams = zod.object({
  "id": zod.uuid()
})

export const GetBookingResponse = zod.object({
  "appointmentDate": zod.iso.date(),
  "appointmentTime": zod.string(),
  "createdAt": zod.iso.datetime({"offset":true}),
  "customerName": zod.string(),
  "id": zod.uuid(),
  "licencePlate": zod.string(),
  "washProgramme": zod.enum(['BASIC', 'COMFORT', 'PREMIUM'])
})