// frontend/src/api/generated/model/createBookingRequest.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Prototype API
 * OpenAPI spec version: 1
 */
import type { CreateBookingRequestWashProgramme } from './createBookingRequestWashProgramme';

export interface CreateBookingRequest {
  appointmentDate: string;
  /**
     * @minLength 1
     * @pattern ^([01]\d|2[0-3]):[0-5]\d$
     */
  appointmentTime: string;
  /**
     * @minLength 1
     * @maxLength 100
     */
  customerName: string;
  /**
     * @minLength 1
     * @maxLength 15
     */
  licencePlate: string;
  washProgramme: CreateBookingRequestWashProgramme;
}